http://code.google.com/p/explorercanvas/issues/detail?id=61 : Adding VML during stroke() takes a really long time
Reported by mbis...@gmail.com, Dec 19, 2009

An application that animates a canvas suffers greatly whenever stroke() is called. Using a JS profiler
I was able to determine that the great majority of the time spent in stroke() is adding the VML to the
element_.

For an animation that only calls stroke() once per loop, this isn't that big of a deal. However, an
animation that composes the scene with lots of paths and calls to stroke is pretty much
unwatchable due to the problems with adding VML rapidly to a document.

The VML items should be cached into a display list and rendered on a timely basis, rather than
immediately during stroke(). This design would improve the performance of complex animations as
well as reduce the CPU usage for the app itself.

Comment 1 by mbis...@gmail.com, Dec 19, 2009

Attached is a patch with a display list that is rendered at 24 fps.